Change OnVersion underlying type from u64 to u32

+++ src/on_version.rs
@@ -7,5 +7,5 @@ use crate::deserialize::{Deserialize, FileVersion};
 #[derive(Default, Deserialize)]
 pub struct OnVersion {
     //TODO: process raw version
-    pub raw_version: u64,
+    pub raw_version: u32,
 }
